Add a service layer between client and backend.
Introduce an additional abstraction (service.js) that takes care of
communication between the JS client and the Java server. The main reason
to introduce it is to allow mocking the service - this way we can run a
frontend app locally, even if the Java backend is down.

The *real* reason is that I haven't managed to start the Google App
Engine, because I forgot my security key, so I'm hacking things to
unblock myself.